requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-login
pandas
sklearn
requests
flask-migrate
email_validator
pillow
